# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: Документация по API к маркетплейсу
  title: Diary API
  version: v1
servers:
- url: http://localhost:8000/
security:
- Basic: []
paths:
  /ads/:
    get:
      operationId: Список всех объявлений
      parameters:
      - in: query
        name: title
        schema:
          type: string
      - description: A page number within the paginated result set.
        in: query
        name: page
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: ""
      tags:
      - Объявления
      x-accepts: application/json
    post:
      operationId: Создать объявление
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ad'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'
          description: ""
      tags:
      - Объявления
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /ads/me/:
    get:
      operationId: Мои объявления
      parameters:
      - in: query
        name: title
        schema:
          type: string
      - description: A page number within the paginated result set.
        in: query
        name: page
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: ""
      tags:
      - Объявления
      x-accepts: application/json
  /ads/{ad_pk}/comments/:
    get:
      operationId: Список отзывов к объявлению
      parameters:
      - in: path
        name: ad_pk
        required: true
        schema:
          type: string
      - description: A page number within the paginated result set.
        in: query
        name: page
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
          description: ""
      tags:
      - Объявления
      x-accepts: application/json
    post:
      operationId: Оставить отзыв
      parameters:
      - in: path
        name: ad_pk
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ""
      tags:
      - Объявления
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /ads/{ad_pk}/comments/{id}/:
    delete:
      operationId: Удалить отзыв
      parameters:
      - in: path
        name: ad_pk
        required: true
        schema:
          type: string
      - description: A unique integer value identifying this Комментарий.
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          content: {}
          description: ""
      tags:
      - Объявления
      x-accepts: application/json
    get:
      operationId: Получить отзыв по ID
      parameters:
      - in: path
        name: ad_pk
        required: true
        schema:
          type: string
      - description: A unique integer value identifying this Комментарий.
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ""
      tags:
      - Объявления
      x-accepts: application/json
    patch:
      operationId: Изменить отзыв
      parameters:
      - in: path
        name: ad_pk
        required: true
        schema:
          type: string
      - description: A unique integer value identifying this Комментарий.
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ""
      tags:
      - Объявления
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /ads/{id}/:
    delete:
      operationId: удалить объявление
      parameters:
      - description: A unique integer value identifying this Объявление.
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          content: {}
          description: ""
      tags:
      - Объявления
      x-accepts: application/json
    get:
      operationId: Подробно об объявлении
      parameters:
      - description: A unique integer value identifying this Объявление.
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdDetail'
          description: ""
      tags:
      - Объявления
      x-accepts: application/json
    patch:
      operationId: Изменить объявление
      parameters:
      - description: A unique integer value identifying this Объявление.
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ad'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'
          description: ""
      tags:
      - Объявления
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /refresh/:
    post:
      description: |-
        Takes a refresh type JSON web token and returns an access type JSON web
        token if the refresh token is valid.
      operationId: обновить токен
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ""
      tags:
      - Аутентификация
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /token/:
    post:
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      operationId: Получить токен
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenObtainPair'
          description: ""
      tags:
      - Аутентификация
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /users/:
    get:
      operationId: Список пользователей
      parameters:
      - description: A page number within the paginated result set.
        in: query
        name: page
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
          description: ""
      tags:
      - Пользователь
      x-accepts: application/json
    post:
      operationId: Регистрация пользователя
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistration'
          description: ""
      tags:
      - Пользователь
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /users/me/:
    get:
      operationId: Мой профиль
      parameters:
      - description: A page number within the paginated result set.
        in: query
        name: page
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
          description: ""
      tags:
      - Пользователь
      x-accepts: application/json
    patch:
      operationId: Изменить мой профиль
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ""
      tags:
      - Пользователь
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /users/reset_password/:
    post:
      operationId: Сброс пароля по email (не обязательно), djoser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendEmailReset'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendEmailReset'
          description: ""
      tags:
      - Пользователь
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /users/reset_password_confirm/:
    post:
      operationId: Задать новый пароль по ссылке с почты, djoser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetConfirm'
          description: ""
      tags:
      - Пользователь
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /users/set_password/:
    post:
      operationId: Изменить пароль
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPassword'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetPassword'
          description: ""
      tags:
      - Пользователь
      x-codegen-request-body-name: data
      x-contentType: application/json
      x-accepts: application/json
  /users/{id}/:
    get:
      operationId: Профиль пользователя по ID
      parameters:
      - description: A unique integer value identifying this Пользователь.
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ""
      tags:
      - Пользователь
      x-accepts: application/json
components:
  schemas:
    Ad:
      example:
        image: https://openapi-generator.tech
        price: 314780940
        author: 5
        pk: 6
        title: title
      properties:
        pk:
          readOnly: true
          title: ID
          type: integer
        image:
          description: Разместите фото для объявления
          format: uri
          nullable: true
          readOnly: true
          title: Фото
          type: string
        title:
          description: введите название товара
          maxLength: 200
          minLength: 1
          title: Название товара
          type: string
        price:
          description: Добавьте цену товара
          maximum: 2147483647
          minimum: 0
          title: Цена товара
          type: integer
        author:
          description: Выберите автора объявления
          title: Автор объявления
          type: integer
      required:
      - author
      - price
      - title
      type: object
    Comment:
      example:
        author: 1
        created_at: 2000-01-23T04:56:07.000+00:00
        pk: 6
        text: text
      properties:
        pk:
          readOnly: true
          title: ID
          type: integer
        text:
          description: Оставьте свой комментарий здесь
          maxLength: 1000
          minLength: 1
          title: Комментарий
          type: string
        author:
          description: Выберите автора комментария
          title: Автор комментария
          type: integer
        created_at:
          description: Введите время создания комментария
          format: date-time
          readOnly: true
          title: Время создания комментария
          type: string
      required:
      - author
      - text
      type: object
    AdDetail:
      example:
        author_first_name: author_first_name
        image: https://openapi-generator.tech
        author_last_name: author_last_name
        phone: phone
        price: 1294386358
        pk: 0
        title: title
      properties:
        pk:
          readOnly: true
          title: ID
          type: integer
        image:
          description: Разместите фото для объявления
          format: uri
          nullable: true
          readOnly: true
          title: Фото
          type: string
        title:
          description: введите название товара
          maxLength: 200
          minLength: 1
          title: Название товара
          type: string
        price:
          description: Добавьте цену товара
          maximum: 2147483647
          minimum: 0
          title: Цена товара
          type: integer
        phone:
          minLength: 1
          title: Phone
          type: string
        author_first_name:
          readOnly: true
          title: Author first name
          type: string
        author_last_name:
          readOnly: true
          title: Author last name
          type: string
      required:
      - phone
      - price
      - title
      type: object
    TokenRefresh:
      example:
        access: access
        refresh: refresh
      properties:
        refresh:
          minLength: 1
          title: Refresh
          type: string
        access:
          readOnly: true
          title: Access
          type: string
      required:
      - refresh
      type: object
    TokenObtainPair:
      example:
        password: password
        email: email
      properties:
        email:
          minLength: 1
          title: Email
          type: string
        password:
          minLength: 1
          title: Password
          type: string
      required:
      - email
      - password
      type: object
    User:
      example:
        phone: phone
        last_name: last_name
        id: 6
        first_name: first_name
        email: email
      properties:
        first_name:
          description: Введите имя, макс 64 символа
          maxLength: 64
          minLength: 1
          title: Имя
          type: string
        last_name:
          description: Введите фамилию, макс 64 символа
          maxLength: 64
          minLength: 1
          title: Фамилия
          type: string
        phone:
          description: Укажите телефон для связи
          maxLength: 128
          minLength: 1
          title: Телефон для связи
          type: string
        id:
          readOnly: true
          title: ID
          type: integer
        email:
          description: Укажите электронную почту
          format: email
          minLength: 1
          readOnly: true
          title: Email address
          type: string
      required:
      - first_name
      - last_name
      - phone
      type: object
    UserRegistration:
      example:
        password: password
        phone: phone
        last_name: last_name
        first_name: first_name
        email: email
      properties:
        email:
          description: Укажите электронную почту
          format: email
          maxLength: 254
          minLength: 1
          title: Email address
          type: string
        first_name:
          description: Введите имя, макс 64 символа
          maxLength: 64
          minLength: 1
          title: Имя
          type: string
        last_name:
          description: Введите фамилию, макс 64 символа
          maxLength: 64
          minLength: 1
          title: Фамилия
          type: string
        password:
          minLength: 1
          title: Password
          type: string
        phone:
          description: Укажите телефон для связи
          maxLength: 128
          minLength: 1
          title: Телефон для связи
          type: string
      required:
      - email
      - first_name
      - last_name
      - password
      - phone
      type: object
    Activation:
      properties:
        uid:
          minLength: 1
          title: Uid
          type: string
        token:
          minLength: 1
          title: Token
          type: string
      required:
      - token
      - uid
      type: object
    SendEmailReset:
      example:
        email: email
      properties:
        email:
          format: email
          minLength: 1
          title: Email
          type: string
      required:
      - email
      type: object
    UsernameResetConfirm:
      properties:
        new_email:
          description: Укажите электронную почту
          format: email
          maxLength: 254
          minLength: 1
          title: Email address
          type: string
      required:
      - new_email
      type: object
    PasswordResetConfirm:
      example:
        uid: uid
        new_password: new_password
        token: token
      properties:
        uid:
          minLength: 1
          title: Uid
          type: string
        token:
          minLength: 1
          title: Token
          type: string
        new_password:
          minLength: 1
          title: New password
          type: string
      required:
      - new_password
      - token
      - uid
      type: object
    SetUsername:
      properties:
        current_password:
          minLength: 1
          title: Current password
          type: string
        new_email:
          description: Укажите электронную почту
          format: email
          maxLength: 254
          minLength: 1
          title: Email address
          type: string
      required:
      - current_password
      - new_email
      type: object
    SetPassword:
      example:
        new_password: new_password
        current_password: current_password
      properties:
        new_password:
          minLength: 1
          title: New password
          type: string
        current_password:
          minLength: 1
          title: Current password
          type: string
      required:
      - current_password
      - new_password
      type: object
    inline_response_200:
      example:
        next: https://openapi-generator.tech
        previous: https://openapi-generator.tech
        count: 0
        results:
        - image: https://openapi-generator.tech
          price: 314780940
          author: 5
          pk: 6
          title: title
        - image: https://openapi-generator.tech
          price: 314780940
          author: 5
          pk: 6
          title: title
      properties:
        count:
          type: integer
        next:
          format: uri
          nullable: true
          type: string
        previous:
          format: uri
          nullable: true
          type: string
        results:
          items:
            $ref: '#/components/schemas/Ad'
          type: array
      required:
      - count
      - results
      type: object
    inline_response_200_1:
      example:
        next: https://openapi-generator.tech
        previous: https://openapi-generator.tech
        count: 0
        results:
        - author: 1
          created_at: 2000-01-23T04:56:07.000+00:00
          pk: 6
          text: text
        - author: 1
          created_at: 2000-01-23T04:56:07.000+00:00
          pk: 6
          text: text
      properties:
        count:
          type: integer
        next:
          format: uri
          nullable: true
          type: string
        previous:
          format: uri
          nullable: true
          type: string
        results:
          items:
            $ref: '#/components/schemas/Comment'
          type: array
      required:
      - count
      - results
      type: object
    inline_response_200_2:
      example:
        next: https://openapi-generator.tech
        previous: https://openapi-generator.tech
        count: 0
        results:
        - phone: phone
          last_name: last_name
          id: 6
          first_name: first_name
          email: email
        - phone: phone
          last_name: last_name
          id: 6
          first_name: first_name
          email: email
      properties:
        count:
          type: integer
        next:
          format: uri
          nullable: true
          type: string
        previous:
          format: uri
          nullable: true
          type: string
        results:
          items:
            $ref: '#/components/schemas/User'
          type: array
      required:
      - count
      - results
      type: object
  securitySchemes:
    Basic:
      scheme: basic
      type: http
x-original-swagger-version: "2.0"